+++ src/cli/index.js
@@ -312,7 +312,10 @@ async function runCommand_db() {
 	await fs.emptyDir(roots.eyasDist);
 
 	// put the user's test into an asar file with .eyas extension
-	const artifactName = `${config.title} - ${config.version}.eyas`;
+	const artifactName = `${config.title} - ${config.version}.eyas`
+		// remove any characters that could cause issues
+		.replace(/[\/\\:\*\?"<>\|]/g, `_`);
+
 	const testSourceDirectory = config.source;
 	const outputSourceDirectory = path.join(roots.eyasDist, `source`);
 	const destinationAsarPath = path.join(roots.eyasDist, artifactName);
@@ -390,7 +393,9 @@ async function runCommand_bundle() {
 		// start a promise for this platform archive process
 		archivePromises.push(
 			new Promise((resolve, reject) => {
-				const artifactName = `${config.title} - ${config.version}.${platform.tag}.zip`;
+				const artifactName = `${config.title} - ${config.version}.${platform.tag}.zip`
+					// remove any characters that could cause issues
+					.replace(/[\/\\:\*\?"<>\|]/g, `_`);
 
 				// create the zip file
 				const output = fs.createWriteStream(path.join(roots.eyasDist, artifactName));