+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add WebdriverIO test runner
 - Set the UI layer to always have blurred background when set as visible
 - Fix issue displaying the UI layer without controls when the window is resized, making Eyas appear unresponsive.
+- Make the file name safe when creating the artifact name
 
 # v3.2.2